pyproject.toml
[build-system]
requires = ["poetry==1.0.0.b4"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.poetry]
authors = ["James Roeder <jmaroeder@gmail.com>"]
description = "API"
name = "api"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = "3.8.0"

dj-database-url = "^0.5.0"
django = {version = "^3.0"}
djangorestframework = "^3.10.3"
django-split-settings = "^1.0.0"
environ-config = "^19.1.0"
gunicorn = "^20.0.0"
poetry = "1.0.0.b4"
psycopg2 = "^2.8.4"

[tool.poetry.dev-dependencies]
# django debugging
django-querycount = "^0.7.0"
nplusone = "^1.0.0"

# static analysis
black = "^19.10b0"
django-stubs = "^1.2.0"
flake8 = "^3.7.9"
flake8-django = "^0.0.4"
flake8-pytest = "^1.3"
flake8-pytest-style = "^0.1.3"
mypy = "^0.740"
pytest = "^5.2.2"
pytest-cov = "^2.8.1"
pytest-deadfixtures = "^2.1.0"
pytest-django = "^3.7.0"
pytest-randomly = "^3.1.0"
pytest-testmon = "^0.9.19"
pytest-timeout = "^1.3.3"
pytest-watch = "^4.2.0"
wemake-python-styleguide = "^0.12.5"
xenon = "^0.5.5"
yamllint = "^1.19.0"

# scratch
ipython = "^7.9.0"
